pci: extract address code to its subpkg
In order to maximize the code reuse, we move the code which handles
PCI addresses to its own subpackage.
This is driven by a upcoming patch, which wants to parse PCI addresses
from strings, without consuming the pci package, to avoid circular
imports.

Only trivial code movement + needed renames
e.g.: AddressFromString() -> address.FromString()
